+++ izanagi_kit/src/lib.rs
@@ -131,7 +131,7 @@ pub use easing::{
 };
 pub use entity::{Entity, EntityAllocator};
 pub use fixed::Fixed;
-pub use fov::{can_see, compute_fov, compute_fov_dist, fov_to_vec};
+pub use fov::{can_see, compute_fov, compute_fov_dist, fov_count_filtered, fov_to_vec};
 pub use fsm::Fsm;
 pub use geometry::{
     chebyshev_distance, diamond, line, line_of_sight, manhattan_distance, rect_contains,
@@ -150,9 +150,9 @@ pub use menu::{Menu, MenuItem};
 pub use msglog::MsgLog;
 pub use multimap::{Connector, MultiMap};
 pub use noise::{
-    fbm_1d, fbm_1d_wrap, fbm_2d, fbm_2d_wrap, fbm_3d, hash_1d, hash_2d, hash_3d, noise_3d_in_range,
-    normalize_noise, ridge_noise_2d, value_noise_1d, value_noise_1d_wrap, value_noise_2d,
-    value_noise_2d_wrap, value_noise_3d,
+    fbm_1d, fbm_1d_wrap, fbm_2d, fbm_2d_in_range, fbm_2d_wrap, fbm_3d, hash_1d, hash_2d, hash_3d,
+    noise_3d_in_range, normalize_noise, ridge_noise_2d, value_noise_1d, value_noise_1d_wrap,
+    value_noise_2d, value_noise_2d_wrap, value_noise_3d,
 };
 pub use parser::{error_count, parse, warning_count};
 pub use passability::PassabilityGrid;